Handle merges properly by finding transactions from related tickets.
This relies on the fact that the effectiveid is changed for a ticket
once merged. We use a custom SQL query to find all transactions
including those from tickets merged in to the current one. This
presents it as a single ticket in Redmine which is neater than the
duplicates approach.

Also skip tickets which have been merged in to another, because
they'll be dealt with by the above SQL query when the ticket they've
been merged in to has been processed.